Transferring DB calls
I changed buildLoginWidget in widgetLogin to just pull in the HTML from
widgetLogin.html, and updated code in a few other places that was
affected (for instance, there was a call to createDebug which assumed
that the “Show Debug” button existed. The call to widgetLogin.html is
asynchronous, so I had to make sure the call to createDebug wasn’t
triggered until the login header was finished.

I have also created the new nodeFunctions class. (We talked about
calling it just node, but I thought it would cause confusion if I
created a new file and called it node.js.) It currently contains two
“big functions” and a helper function. The first function is
specifically for logging in, as we talked about. It’s commented out
because I think I had a better idea, but I didn’t want to delete it in
case I had missed something.

The second function in the nodeFunctions class is called
changeRelation, and it is more general. It can find, update and return
any relation. It takes four arguments - a data object, which contains
objects to describe the relation, the nodes at either end, and any
changes to make; and a callback object, method and arguments list. At
the moment, logging in calls this function. There are probably four or
five other places in the code where DB calls could also be replaced by
a call to this function.

I have successfully logged into the Harmony site using node.js on my PC
using both the specific login function and the changeRelation function.