Rewriting DB queries
Rewrote all DB queries to call XHTTP requests in the new nodeFunctions
class, rather than writing the queries themselves and calling the old
DB class. The nodeFunctions class ended up with 13 functions, not
counting the empty constructor: 2 helper functions; 6 basic functions
for creating, reading/changing and deleting nodes and relations; 2 more
advanced generic functions for finding a node with an optional relation
and finding a pattern of three connected nodes; and three specific
functions for queries that I was not able to easily combine with
anything else.

The site SHOULD now work when running on a server. I tested each change
as I made it on the Mac (running the DB locally). When I finished, I
tested a few of the most obvious functions (opening and changing a
mindmap, opening the people table and then opening a node from it,
etc.) on the PC (using the Mac as a server), but it was fairly late and
I did not test on the PC extensively. There are also a lot of comments
I need to clean up - I commented out the old code, but didn’t remove it
in case I needed to refer back to it.