Cleaning up code; small bug fix, adjustment to mindmap
I made some behind-the-scenes changes to how the nodeFunctions methods
work, which I hope will make them easier to use when we write new
functions that work with them - that shouldn’t have affected how the
program currently runs. I found and fixed a bug that occurred when
saving an SVG file - the name and number of labels weren’t being stored
properly. I adjusted dragging in mind maps so that the code to
determine whether one label gets attached to another doesn’t look at
the mouse position anymore. For now, it uses the same code as it did
for the mouse position, but checks the center of the left edge of the
node being dragged. I think that should have similar functionality to
what we discussed, and only required changing two lines of code instead
of rewriting several methods. But if this won’t work, let me know and
I’ll redo it.