Minor changes as discussed yesterday
I made most of the changes we talked about yesterday:

I moved the helper functions in nodeFunctions.js to the bottom of the
class, so that it would be easier for others to find the functions they
needed to know about.

I rewrote the code to detect when one label in a mindmap should be
linked to another. Now, it checks whether the CENTER of the label being
dragged is on top of another label, whether the CENTER of the label
being dragged is just above or below another label, and whether the
LEFT EDGE of the label being dragged is to the right of another label.
I think this should work well, but we’ll keep playing with it to make
sure.

I deleted the db class, which is no longer needed.

I created a new app.error method which takes an error message as an
argument. At the moment, it alerts that error message and the stack.
The stack includes the line number and function name at each stage, so
it should provide some useful information. I changed several places in
the code which alerted error messages to send the message to app.error
instead. However, I did not change alerts which seemed clearly meant to
be seen by the user (like “Enter your username and password before
logging in”) or which were not error messages (like the one in
widgetRegression which tells you the name of the file being played).