New branch with updated saving and deleting
I created this new branch because I was excited about getting to work,
but the changes I wanted to make were too extensive to mess around in
an existing branch without permission. Please look it over and let me
know whether I should merge it.

Changes:

When data are deleted, they now turn red instead of disappearing right
away. For the time being, you can select deleted data and move it
around, you can detach it from its parent, and you can drag it to
non-deleted data to make it a child. You canNOT make other labels
(deleted or not) the children of deleted labels, or connect nodes,
files or URLs to deleted labels.

When you hover over a deleted label, instead of the usual “details”,
“edit”, “notes” and “children” toggles, a “Restore” button is
displayed. Clicking this button removes the red highlighting from that
node and all its descendants. If the node’s parent is deleted,
restoring it detaches it from its parent.

When you save the mind-map, deleted data are removed permanently.

Also, mindmaps now have links to all nodes which are linked to their
labels. Every time you save a mindmap, links are created or deleted as
needed. When a mindmap is loaded, the nodes it is linked to are
included in the data, and any changes made to them appear in the
mindmap. Older mindmaps should still be able to load, and should
automatically create these links when they are saved for the first time.

Mindmaps also have a link to their owner - the person who created them.
This is not used for anything yet, but will come in handy when we start
adding permissions (which only the owner can set) and restricting which
mindmaps appear in a search (it should be only the ones that the
logged-in user owns or has permission to see). When you create a new
mindmap, save a new copy of a mindmap, or save an old mindmap which did
not have an owner, you become its owner.

A few unrelated minor updates - I found an error in one of the
nodeFunctions functions and fixed it; found a place where I hadn’t
accounted for a change made when I updated the highlightParent function
and changed that; and made mindmaps create a new node rather than merge
when another mindmap with the same name exists (to make them more
consistent with other nodes).

Some things I want to discuss:

Are current interactions with deleted data OK? Is there more users
should be able to do with it? Anything they SHOULDN’T be able to do?

When the user restores a deleted label, and its parent is deleted, is
it OK to detach the child? Other options I can see are to restore all
ancestors too or to move it so it’s attached to its first ancestor that
wasn’t deleted (if any). Just leaving it as the child of deleted data
doesn’t seem like a reasonable idea to me.