Enabled searching by ownership and permissions
In the "people" table, there is now a dropdown list allowing the user to select "All People", "Users and Admins", "Users" or "Admins". In the "mindmap" table, there is now a textbox and dropdown list allowing the user to search for mindmaps belonging to a specific person - it defaults to the user.