Increased customization; ability to copy
Users can now change the label they see for a node type (“person”
instead of “people”, for example), and the label for a node field
(“Full Name” instead of “Name”, for example). They can rearrange fields
in a node form, which changes the order in which the fields are
displayed on both tables and forms. They can choose which fields to
have displayed on tables, and which to show on forms by default.

Also, a new page has been added to the main menu - “Copy to another
database”. That leads to a page with an input, a start button and a
progress table. Type the IP address or url of the destination (with an
empty DB running) into the input, click the button and it will copy
everything to the other DB and show its progress as it works.