Improvements to backup; removed GUIDs from display
The GUIDs are too big to be usefully displayed, so I hid the areas in
widgetRelations and widgetTableNodes which showed them. Rows in
widgetTableNodes can now be edited by clicking on the row number. I
changed the backup system so that it no longer stores Neo4j IDs of any
kind, it always stores the nodes and relations in the same order, and
it always stores properties in the same order. It should therefore give
identical results when used twice on the same database.