Moved CRUD functions to server
I moved all the functions from the nodeFunctions class onto the server.
They can all be called with an xhttp request where server = "CRUD",
function = the name of the function to call, query = the data object
which describes the desired query, and (for the tableNodeSearch,
addNodeToView and getMetaData functions, whose outputs depend in part
on the user ID) ID = the user's ID.

I removed the uuid option for the server, and the function it led to
(which generated a specified number of UUIDs and sent them to the
client), because that code was only needed in order for the client-side
CRUD functions to generate UUIDs, and there are no more client-side
CRUD functions.

I removed all references to the old nodeFunctions class, which is no
longer used, but I haven't deleted it yet because I might want to reuse
some code I had left in there from functions I hadn't finished writing.
I should be able to salvage all I need and delete the rest within a day
or two.