Introduced changelogs, removed merge command
The client can still send the server a merge request, but the server
now handles it by searching, creating and/or updating rather than using
the merge command.

Whenever a node or relationship is created, changed or deleted, a
changeLog node is created which logs the GUID of the node, the change
which was made, and the GUID of the user who made it.