Change table UI
There can now be only one open search table at a time, and it's always
at the top of the page. When you open a new one, the one that was open
(if any) is hidden. Also, fixed a bug in which if a node had a label
different from its name in the DB, that node type couldn't be opened
from a mindmap. Also rewrote the code for most of the calls to CRUD
functions in a more compact and readable form.