Enabled dragging from node form to view or mindmap
I created a rough approximation of this feature. It's a piece of bold
text saying, "Drag To View". If you click on it and drag, you can drag
the node to a view or a mindmap, just as if you were dragging from a
widgetTableNodes widget. I still need to consider a better design, and
there's a small inconsistency where dragging from a node to a mindmap
stores more details than dragging from a table, but it's a start.