Include direct links between nodes linked by views; include summary button; fixed some GUID/ID bugs
When a user adds a relation to their view of a node, the code now
creates or updates a link between the two nodes the user believes are
related. The link has the label "directRelation" and stores a count,
which is the number of times those two nodes have been linked by users'
views - every time the another user adds this node to the other node's
view, the count is incremented. When a user DELETES a node from their
view, the relation's count is decremented, or if it was already 1
(meaning that the link being deleted was the only one), it is deleted.

At the top of the relations widget, next to the refresh button, there
is now a "summary" button which shows all the nodes which are linked to
the subject node, sorted by the number of times they are linked.

While working on this, I found and fixed some bugs in which IDs and
GUIDs were being used inconsistently (i.e., I was searching for a node
as if its GUID were its Neo4j ID). I suspect there are more that I
haven't found.