Created integrity module
The new integrity module is run by the server. It exports a class,
"integrity". Instances of this class can, at the moment, check to
ensure that all nodes and relations have GUIDs and that the directLinks
(which record how many times two nodes are linked through users' views)
accurately track the Links. These three functions can be called
together by integrity.all(), or separately by
integrity.missingNodeGUIDS(), integrity.missingRelGUIDS(), and
integrity.findLinks(). The integrity constructor takes two arguments -
the driver to use for database queries, and the uuid generator
function. There is one other instance variable, a boolean called
"update", which defaults to false. If it is false, the integrity object
will alert the user to problems in the database, but will not make
changes. If it is true, the integrity object will generate new GUIDs
for any node or relation that is missing them, and will create, remove
or update directLinks as necessary to make them accurate.