Enabled dragging nodes to search; minor updates; behind-the-scenes changes
There is now a row of boxes above the div with the search table, into
which you can drag nodes. They can be dragged from a search table, a
view or a node form - all the places you can drag nodes from. When an
empty box is filled with a node, a new empty box appears, so there is
always one empty box. You can replace a node by dragging the new one
over the old one, or remove it by clicking the "X" button in its box.
When a search is performed, only nodes which are linked to ALL nodes
shown in this row will be returned.

Other than that, the only UI changes are that the "Save" button on
metadata nodes (which shouldn't be changeable by the user) has been
removed, and the server code produces a few more messages about the
checks it runs when the server starts up.

Behind the scenes, I created a new "integrity" module for the server
and wrote code to verify that all nodes and relations have GUIDs and
that any two nodes linked by a user's view are also linked by a
directLink relation. I then moved several functions which had been
elsewhere in the code, but were related to database integrity, into
this function: the code to verify that all metadata nodes include their
user-defined labels, the code to verify that the preset calendar(s)
exist, and the code to verify that the login tables exist and there is
at least one admin.

I've also been working on cleaning up code - rewriting parts that were
done inefficiently, removing variables that are never used, making sure
all instance variables are set in the constructor so it's easier to
keep track of what variables exist in a class, etc.