Working on progress bar; fixed bug in Chrome
The progress box now appears at the top right of the screen, says
"Available" when no request is being processed, and shows a description
of the request, the time it's taken so far and a cancel button when a
request is being processed. It can show multiple requests. It does not
yet interact with the database at all, and the cancel button doesn't do
anything yet.

Mary pointed out a bug in which data in node tables was not selectable
in Chrome. I think I've fixed it.