Status bar; records of requests; Data Browser improvements
Finished implementing the first draft of the status bar. It displays a
message describing all requests which are currently active, or
"Available" if none are. Each request has a cancel button which becomes
enabled if more than a second has passed without the request finishing.
The cancel button removes the request from the status bar, but at
present, doesn't do anything else. The status bar does not currently
report errors. Those improvements are planned for the next draft.

Records are now kept in the database of what requests are made to the
database. When a user first attempts to log in, a "M_Session" node is
created, an M_Browser node representing their browser is created if it
didn't already exist, and a relation is created between them
representing the login attempt. If the user successfully logs in, the
session is linked to their node as well. Every request they make to the
DB generates a new link between the session and the browser. The links
record a description of the request, a start time, an end time and an
end condition such as "Succeeded" or "Cancelled". The session itself
stores a start time and an end time, although at the moment, the end
time is not recorded if the user ends the session by closing the page
instead of logging out.

The Data Browser tool has been revamped. It now consists of a row of
five cells, each of which is labeled for easier use. The Main cell, in
the middle, can be dragged to; this displays the node which was dragged
in in the Main cell, its incoming relations on the left and its
outgoing relations on the right. The lists of relations now have a max
height of 150px; if the list is longer, it becomes scrollable.

Clicking a relation displays the node it goes to in the far left or far
right cell, and displays the details of that relation underneath the
list of relations on that side (so clicking a relation on the left
makes the cell appear on the far left, and the relation details appear
under the list of relations on the left). Clicking a different relation
will replace the relation and node details that are already shown with
the ones for that relation; clicking the relation that is already open
will close it again.

There are arrows above each list of relations which are enabled if a
relation is open on that side; clicking the arrow moves the "user's
view" in that direction. For instance, if you click the left arrow, the
node which was on the left becomes the new main node. The old main node
remains open as the new right node.

Popups were getting in the way, and the redesign made the issue worse,
so they have been disabled for now pending better ideas about how to
handle them.