Fixed a couple of bugs; cleaned up some unused variables and old code; extracted two new server modules
Most of the changes I've made are small ones that I noticed as I went
through the code working on documentation. I corrected typos and
outdated comments, replaced complicated code with simpler versions,
removed variables that are never used, etc. But I'm pushing an update
because I found a serious bug in widgetView which, whenever a user
created a new view, was accidentally linking that user to every other
node in the DB. I have fixed this bug now.