Added first draft of help and bug report buttons
Each top-level widget (that is, each widget with its own header) now
has a "?" button and a "!" button. Clicking the "?" button opens a node
describing that type of widget, with a "Help" field that may acquire
useful information later. Clicking the "!" button, for now, opens the
"Harmony Changes" topic, but it will be easy to make it open a node
specific for issues with a given type of node eventually.