Moved node details in mindmap
There is no longer a details popup in mindmap. Instead, when you click
the "Show Details" button, the details pane which opens shows details
of the mindmap if no label, more than one label or a label without
details is selected; otherwise, it shows the details of the selected
label.