Improved logic for combining old and new versions of metadata
Hopefully, the new logic should a) make it possible for the metadata
nodes in the DB to be updated more easily, and b) keep old user
settings from interfering with new features. It basically goes through
the entire metadata object describing a node, AND every property of
every object within it, and all THEIR properties if they're objects,
and so on recursively. At every level, the version from the user's
settings object is used if it exists; if not, the version from the
metadata node is used if it exists; and if not, the version from the
metadata class is used. That means we can add new properties to the
metadata class at any level, and they will NOT be overwritten because
that specific property won't exist anywhere else. If we want to CHANGE
a property, we'll still have to do that manually.