Allowed opening a node from a relation
When looking at the relations of a node, you can now click on any of
the first three visible cells (the row number, name and type) and, if
that row represents a node, the node will open. (If it's just a
comment, nothing will happen.) I also fixed a couple of small bugs -
one in the new app.updateObject method which was failing to call itself
recursively when it was supposed to, and one in widgetTableNodes.edit
which would have caused trouble later if anyone had renamed the mindmap
or calendar nodes.