Finished migrating progress code
Records of database calls are now stored in the header of the widget
they're associated with, except for the initial queries to update
metadata and the checkEmpty method, neither of which IS associated with
a widget. Later plans are to move those initial queries onto the server
and put the checkEmpty button in a widget.