Implemented download button and dropdown for files
At the moment: When you open a file node which has at least one file
associated with it, a dropdown list of the versions and a download
button appear. The dropdown defaults to selecting the latest version of
the file. The download button downloads the version currently selected
in the dropdown. If you create a file node/open a file node with no
files associated with it, and then add a file, the download button does
NOT appear unless the node is closed and reopened. Adding it in that
situation is a low priority because the user just UPLOADED that file in
the current session - it must already exist on their system, and they
must be able to access it.

Also, we have confirmed that the methods I'm currently using to upload
and download files only work on text.