Widget list
There's now a list of widgets in the top right corner (with sections
for header widgets, normal widgets and minimized widgets). You can drag
and drop items in that list to rearrange them; this also rearranges the
actual widgets on the screen. Clicking a minimized widget reopens it
and, and clicking any widget scrolls to it.