Implemented token login
First draft of token login. The login.js server module includes a
"tokens" object. Its keys are session GUIDs and its values are objects
describing the logged-in user (their name, GUID, etc.). Tokens expire
after 30 minutes of inactivity. When the user sends a CRUD request,
they also send their session GUID as a token. First the server checks
whether that session GUID is in the list of tokens; then, if it is, the
server resets the timer for that token and processes the request. If
the token is NOT in the active list, the server sends back the word
"timeout", and the client alerts the user that their session has ended
and logs them out.