Finished making dummy pages for website
Each page contains a header, footer, navigation menu and a sentence of
dummy content. The only changes from one page to another are which item
in the nav menu has the "current-menu-item" class (which marks it as
selected) and what the dummy content says. Hope to extract this into a
template eventually. I also copied all the images that the site
referenced into a local folder. There are no longer any references to
etpri.org on this copy site.